exceptions za rating

+++ rating-service/src/main/java/com/example/ratingservice/servisi/StripServis.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
+import com.example.ratingservice.exception.ApiRequestException;
 import com.example.ratingservice.modeli.Korisnik;
 import com.example.ratingservice.modeli.Strip;
 import com.example.ratingservice.repozitorij.StripRepozitorij;
@@ -17,22 +18,31 @@ public class StripServis {
 	@Autowired
 	private StripRepozitorij stripRepozitorij;
 	
-	public Optional<Strip> findById(Long id) {
-		return stripRepozitorij.findById(id);
-	}
 	public Strip getOne(Long id) {
-		return stripRepozitorij.getOne(id);
-	}
-	public Iterable<Strip> findAll() {
-		return stripRepozitorij.findAll();
+		
+		if(stripRepozitorij.findById(id).isPresent()) {
+			return stripRepozitorij.getOne(id);
+		}
+		
+		throw new ApiRequestException("Strip sa id "+id.toString()+" nije pronađen!");
+		
 	}
 	
 	public void save(Strip strip) {
 		stripRepozitorij.save(strip);
 	}
+	
 	public void deleteById(Long id) {
-		stripRepozitorij.deleteById(id);
+		
+		if(stripRepozitorij.findById(id).isPresent()){
+			stripRepozitorij.deleteById(id);
+		}
+		
+		throw new ApiRequestException("Strip sa id "+id.toString()+" nije pronađen!");
 	}
-
+	
+	public Iterable<Strip> findAll() {
+		return stripRepozitorij.findAll();
+	}
 	
 }
